+++ source/Nuke.Core/OutputSinks/ConsoleOutputSink.cs
@@ -19,7 +19,7 @@ namespace Nuke.Core.OutputSinks
     {
         public virtual void Write(string text)
         {
-            WriteWithColors(text, ConsoleColor.White, ConsoleColor.Black);
+            WriteWithColors(text, ConsoleColor.White);
         }
 
         public virtual IDisposable WriteBlock(string text)
@@ -33,26 +33,26 @@ namespace Nuke.Core.OutputSinks
 
         public virtual void Trace(string text)
         {
-            WriteWithColors(text, ConsoleColor.Gray, ConsoleColor.DarkGray);
+            WriteWithColors(text, ConsoleColor.DarkGray);
         }
 
         public virtual void Info(string text)
         {
-            WriteWithColors(text, ConsoleColor.White, ConsoleColor.Black);
+            WriteWithColors(text, ConsoleColor.White);
         }
 
         public virtual void Warn(string text, string details = null)
         {
-            WriteWithColors(text, ConsoleColor.Yellow, ConsoleColor.DarkYellow);
+            WriteWithColors(text, ConsoleColor.DarkYellow);
             if (details != null)
-                WriteWithColors(details, ConsoleColor.Yellow, ConsoleColor.DarkYellow);
+                WriteWithColors(details, ConsoleColor.DarkYellow);
         }
 
         public virtual void Error(string text, string details = null)
         {
-            WriteWithColors(text, ConsoleColor.Red, ConsoleColor.DarkRed);
+            WriteWithColors(text, ConsoleColor.DarkRed);
             if (details != null)
-                WriteWithColors(details, ConsoleColor.Red, ConsoleColor.DarkRed);
+                WriteWithColors(details, ConsoleColor.DarkRed);
         }
 
         public virtual void WriteSummary(IReadOnlyCollection<TargetDefinition> executionList)
@@ -85,19 +85,13 @@ namespace Nuke.Core.OutputSinks
         }
 
         [MethodImpl(MethodImplOptions.Synchronized)]
-        private void WriteWithColors(string text, ConsoleColor brightForeground, ConsoleColor darkForeground)
+        private void WriteWithColors(string text, ConsoleColor foregroundColor)
         {
-            var previousForeground = Console.ForegroundColor;
-            var backgroundColor = Console.BackgroundColor;
-
-            // TODO: can we determine the actual console background color?
-            var hasDarkBackground = (int) backgroundColor == -1
-                                    || backgroundColor == ConsoleColor.Black
-                                    || backgroundColor.ToString().StartsWith("Dark");
+            var previousForegroundColor = Console.ForegroundColor;
 
             using (DelegateDisposable.CreateBracket(
-                () => Console.ForegroundColor = hasDarkBackground ? brightForeground : darkForeground,
-                () => Console.ForegroundColor = previousForeground))
+                () => Console.ForegroundColor = foregroundColor,
+                () => Console.ForegroundColor = previousForegroundColor))
             {
                 Console.WriteLine(text);
             }